% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench.R
\name{report_comparison}
\alias{report_comparison}
\title{simple comparison of retrieval from Azure/AWS}
\usage{
report_comparison(
  location = "Boston",
  target_file = tempfile(),
  stub = "https://annotationhub.bioconductor.org/",
  cache = BiocFileCache::BiocFileCache(),
  tags = c("5050", "2"),
  hosts = c("azure", "fetch"),
  ntimes = 2
)
}
\arguments{
\item{location}{character(1) description of location of run, please make informative}

\item{target_file}{character(1) file to which report is written in CSV}

\item{stub}{character(1) URL stub to which `fetch/[num]` is appended for retrieval}

\item{cache}{instance of BiocFileCache}

\item{tags}{character() `fetch_id` values from `AnnotationHub::getInfoOnIds`}

\item{hosts}{character() defaults to `c("azure", "fetch")` for use in API call}

\item{ntimes}{numeric() defaults to 2, number of retrievals to be benchmarked}
}
\description{
simple comparison of retrieval from Azure/AWS
}
